[feature/ms-339] delete Price List

+++ apps/masterdata/saledata/serializers/price.py
@@ -484,15 +484,18 @@ class PriceDeleteSerializer(serializers.ModelSerializer):  # noqa
         fields = ()
 
     def update(self, instance, validated_data):
-        if not Price.objects.filter_current(
-                fill__tenant=True,
-                fill__company=True,
-                price_list_mapped=instance.id
-        ).exists():
-            ProductPriceList.objects.filter(price_list=instance).delete()  # delete all item in M2M table
-            instance.delete()  # delete price list
-            return True
-        raise serializers.ValidationError(PriceMsg.PARENT_PRICE_LIST_CANT_BE_DELETE)
+        if ProductPriceList.objects.filter(price_list=instance).exists():
+            raise serializers.ValidationError(PriceMsg.NON_EMPTY_PRICE_LIST_CANT_BE_DELETE)
+        else:
+            if not Price.objects.filter_current(
+                    fill__tenant=True,
+                    fill__company=True,
+                    price_list_mapped=instance.id
+            ).exists():
+                ProductPriceList.objects.filter(price_list=instance).delete()  # delete all item in M2M table
+                instance.delete()  # delete price list
+                return True
+            raise serializers.ValidationError(PriceMsg.PARENT_PRICE_LIST_CANT_BE_DELETE)
 
 
 class PriceListUpdateItemsSerializer(serializers.ModelSerializer):  # noqa
